- No more AbstractTableFilter class needed -> Only JTableFilter now     - I merged the AbstractTableFilter with JTableFilter     - All the code for the JTableFilter is at least on one file now       - there are also a nested Row interface and a RowSorter class in JTableFilter     - AbstractTableFilter is no longer needed -> it was deleted
