--- src/main/java/com/elle/analyster/presentation/filter/AbstractTableFilter.java
@@ -1,173 +1,0 @@
-package com.elle.analyster.presentation.filter;
-
-import javax.swing.*;
-import javax.swing.event.TableModelEvent;
-import javax.swing.event.TableModelListener;
-import javax.swing.table.TableModel;
-import java.beans.PropertyChangeEvent;
-import java.beans.PropertyChangeListener;
-import java.util.*;
-
-/**
- * Partial implementation of table filter
- *
- * Created on Feb 10, 2011
- *
- * @author Eugene Ryzhikov
- *
- * @param <T>
- */
-
-public class AbstractTableFilter {
-
-    private final Set<IFilterChangeListener> listeners = Collections.synchronizedSet(new HashSet<IFilterChangeListener>());
-
-    private final Map<Integer, Collection<DistinctColumnItem>> distinctItemCache
-            = Collections.synchronizedMap(new HashMap<Integer, Collection<DistinctColumnItem>>());
-
-    private JTable table = new JTable(); // this was Type T
-    private TableFilterState filterState = new TableFilterState();
-    
-    /**
-     * CONSTRUCTOR
-     * @param table 
-     */
-    public AbstractTableFilter(JTable table) {
-        this.table = table;
-        setupDistinctItemCacheRefresh();
-    }
-
-    public void setupDistinctItemCacheRefresh() {
-        clearDistinctItemCache();
-        this.table.addPropertyChangeListener("model", new PropertyChangeListener() {
-            @Override
-            public void propertyChange(PropertyChangeEvent e) {
-                clearDistinctItemCache();
-                TableModel model = (TableModel) e.getNewValue();
-                if (model != null) {
-                    model.addTableModelListener(new TableModelListener() {
-
-                        @Override
-                        public void tableChanged(TableModelEvent e) {
-                            clearDistinctItemCache();
-                        }
-                    });
-                }
-            }
-        });
-    }
-
-    private void clearDistinctItemCache() {
-        distinctItemCache.clear();
-    }
-    
-    public void setFilterState(int column, Collection<DistinctColumnItem> values) {
-    filterState.setValues(column, values);
-    }
-    
-    protected boolean execute(int col, Collection<DistinctColumnItem> items){
-        // override in JTable
-        return false;
-    }
-
-    public final void fireFilterChange() {
-        for (IFilterChangeListener l : listeners) {
-            l.filterChanged((JTableFilter) this); 
-        }
-    }
-    
-    private Collection<DistinctColumnItem> collectDistinctColumnItems(int column) {
-        Set<DistinctColumnItem> result = new TreeSet<DistinctColumnItem>(); // to collect unique items
-        for (int row = 0; row < table.getModel().getRowCount(); row++) {
-            Object value = table.getModel().getValueAt(row, column);
-            
-            // handle null exception
-            if(value == null)value = "";
-            
-            result.add(new DistinctColumnItem(value, row));
-        }
-        return result;
-    }
-    
-    /***************************************************************************
-     ********************** Override Methods ***********************************
-     ***************************************************************************/
-    
-    //@Override
-    public JTable getTable() {
-        return table;
-    }
-
-    //@Override
-    public boolean apply(int col, Collection<DistinctColumnItem> items) {
-        setFilterState(col, items); 
-        boolean result = false;
-        if (result = execute(col, items)) {
-            fireFilterChange();
-        }
-        return result;
-    }
-
-    //@Override
-    public boolean apply(int col, Object selectField) { //Create Collection from selected fields 
-        Collection<DistinctColumnItem> item = new ArrayList<>();
-        
-        // handle null exceptions
-        if(selectField == null) selectField = "";
-        
-        DistinctColumnItem distinctColumnItem =new DistinctColumnItem(selectField, col);
-        item.add(distinctColumnItem);
-        return apply(col, item);
-    }
-    //@Override
-    public final void addChangeListener(IFilterChangeListener listener) {
-        if (listener != null) {
-            listeners.add(listener);
-        }
-    }
-
-    //@Override
-    public final void removeChangeListener(IFilterChangeListener listener) {
-        if (listener != null) {
-            listeners.remove(listener);
-        }
-    }
-
-    //@Override
-    public Collection<DistinctColumnItem> getDistinctColumnItems(int column) {
-        Collection<DistinctColumnItem> result = distinctItemCache.get(column);
-        if (result == null) {
-            result = collectDistinctColumnItems(column);
-            distinctItemCache.put(column, result);
-        }
-        return result;
-
-    }
-
-    //@Override
-    public Collection<DistinctColumnItem> getFilterState(int column) {
-        return filterState.getValues(column);
-    }
-
-    //@Override
-    public boolean isFiltered(int column) {
-        Collection<DistinctColumnItem> checks = getFilterState(column);
-        return !(CollectionUtils.isEmpty(checks)) && getDistinctColumnItems(column).size() != checks.size();
-    }
-
-    //@Override
-    public boolean includeRow(JTableFilter.Row row) {
-        return filterState.include(row);
-    }
-
-    //@Override
-    public void clear() {
-        filterState.clear();
-        Collection<DistinctColumnItem> items = Collections.emptyList();
-        for (int column = 0; column < table.getModel().getColumnCount(); column++) {
-            execute(column, items);
-        }
-        fireFilterChange();
-    }
-
-}
